added collection `15` and fixed validator
