--- src/database/entities/data.ts
@@ -1,17 +1,0 @@
-export class DataPayload {
-  id: string;
-
-  filename: string;
-
-  parsed_image_dir: string;
-
-  team_id: number;
-
-  hcp_ori_document_dir: string;
-
-  file_type: string;
-
-  file_page_size: number;
-
-  original_filename: string;
-}
